--- tests/CompositionTestClass.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Blabs\Dictionary\Tests;
-
-use Blabs\Dictionary\WithDictionary;
-
-class CompositionTestClass
-{
-    use WithDictionary;
-
-    const AUBERGINE = 'aubergine';
-    const SWEET_PEPPER = 'sweet pepper';
-    const TOMATO = 'tomato';
-}
